Block manipulations with superuser roles when using set_user_u
This commit introduces new functionality and new GUC variable,
`set_user.block_superuser_manipulation`

When it is enabled, extention will block all attempts to create a new
superuser roles and change/drop existing superuser roles. Also it will
block attemts to grant ROLE with SUPERUSER attribute to some other role.

Before this patch it was possible to create a new superuser, login with
it and do something destructive. All this actions will not be logged.